+++ stable/sclrgstry.cpp
@@ -529,9 +529,21 @@ qRE
 bso::bool__ sclrgstry::BGetValue(
 	const registry_ &Registry,
 	const rgstry::tentry__ &Entry,
+	eNeedness Needness,
 	str::string_ &Value )
 {
-	return Registry.GetValue( Entry, Value );
+	bso::sBool Return = false;
+
+	switch ( Needness ) {
+	case nMandatory:
+		MGetValue( Registry, Entry, Value );
+		break;
+	case nOptional:
+		Return =  OGetValue( Registry, Entry, Value );
+		break;
+	}
+
+	return Return;
 }
 
 void sclrgstry::AddValue(
@@ -581,7 +593,7 @@ bso:: bool__ sclrgstry::OGetValue(
 	const rgstry::tentry__ &Entry,
 	str::string_ &Value )
 {
-	return BGetValue( Registry, Entry, Value );
+	return Registry.GetValue( Entry, Value );
 }
 
 const char *sclrgstry::OGetValue(